refactor: move phase models into main agent file
- Move Pydantic models from separate phase_models.py into run_manager_developer_agents.py
- Delete phase_models.py file as requested in GitHub comment
- Update imports to use pydantic.BaseModel directly
- Maintain unified script interface with no separate code paths
- All functionality preserved including handoffs mechanism
